requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2
django-browser-reload==1.18.0
django-tailwind==4.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
